feat: 커스텀훅으로 도메인 로직 분리
