+++ README.md
@@ -3,7 +3,7 @@
 ### 🎯 핵심 기능
 
 - [] 카드 번호 입력 UI 구성 (4개 필드, 각 4자리)
-- [ ] 각 필드 입력 완료 시 다음 필드로 자동 포커싱
+- [x] 각 필드 입력 완료 시 다음 필드로 자동 포커싱
 - [ ] 숫자 외 입력 시 에러 처리
 - [ ] 카드사 드롭다운 구현
   - [ ] 카드사 선택 시 색상 적용
