+++ src/App.tsx
@@ -5,24 +5,23 @@ import CardNumberSection from './components/CardNumberSection/CardNumberSection'
 import CardExpirationSection from './components/CardExpirationSection/CardExpirationSection';
 import Card from './components/Card/Card';
 import CvcSection from './components/CvcSection/CvcSection';
-import { CvcType } from './types';
-import { INITIAL_CVC } from './constants';
 import CardCompanySection from './components/CardCompanySection/CardCompanySection';
 import useExpiration from './hooks/useExpiration';
 import useCardNumbers from './hooks/useCardNumbers';
+import useCvc from './hooks/useCvc';
 
 export default function App() {
-  const [cvc, setCvc] = useState<CvcType>(INITIAL_CVC);
   const [company, setCompany] = useState<string>('');
 
   const handleSelect = (value: string) => {
     setCompany(value);
   };
 
-  const { expiration, handleExpirationChange, ref } = useExpiration();
+  const { expiration, handleExpirationChange, inputRef } = useExpiration();
 
-  const { inputRefs, handleCardNumberChange, cardNumbers, getCardNumberErrorMessage } = useCardNumbers();
+  const { cardNumbers, handleCardNumberChange, inputRefs, getCardNumberErrorMessage } = useCardNumbers();
 
+  const { cvc, handleCvcChange } = useCvc();
   return (
     <div className={styles.appContainer}>
       <Card numbers={cardNumbers} company={company} expiration={expiration} />
@@ -33,8 +32,8 @@ export default function App() {
         inputRefs={inputRefs}
         getCardNumberErrorMessage={getCardNumberErrorMessage}
       />
-      <CardExpirationSection expiration={expiration} onExpirationChange={handleExpirationChange} ref={ref} />
-      <CvcSection cvc={cvc} setCvc={setCvc} />
+      <CardExpirationSection expiration={expiration} onExpirationChange={handleExpirationChange} ref={inputRef} />
+      <CvcSection cvc={cvc} handleCvcChange={handleCvcChange} />
     </div>
   );
 }
