+++ src/components/CardCompanySection/CardCompanySection.tsx
@@ -1,6 +1,6 @@
+import { CARD_COMPANY } from '../../constants';
 import Dropdown2 from '../Dropdown/Dropdown2';
 import { InputSection } from '../InputSection/InputSection';
-const CARD_COMPANY = ['BC카드', '신한카드', '카카오뱅크', '현대카드', '우리카드', '롯데카드', '하나카드', '국민카드'];
 
 export default function CardCompanySection({ value, onSelect }: { value: string; onSelect: (value: string) => void }) {
   return (
