+++ src/components/CvcSection/CvcSection.tsx
@@ -2,19 +2,13 @@ import { CvcType } from '../../types';
 import Input from '../Input/Input';
 import { InputSection } from '../InputSection/InputSection';
 import styles from './CvcSection.module.css';
-import { Dispatch, SetStateAction } from 'react';
 
 type Props = {
   cvc: CvcType;
-  setCvc: Dispatch<SetStateAction<CvcType>>;
+  handleCvcChange: (value: string) => void;
 };
 
-export default function CvcSection({ cvc, setCvc }: Props) {
-  const handleCvcChange = (value: string) => {
-    const errorMessage = getCvcErrorMessage(value);
-    setCvc({ errorMessage, value });
-  };
-
+export default function CvcSection({ cvc, handleCvcChange }: Props) {
   return (
     <div className={styles.sectionContainer}>
       <InputSection.TitleWrapper>
@@ -30,13 +24,3 @@ export default function CvcSection({ cvc, setCvc }: Props) {
     </div>
   );
 }
-
-const getCvcErrorMessage = (value: string) => {
-  if (!/^[0-9]*$/.test(value)) {
-    return '숫자만 입력 가능합니다.';
-  }
-  if (value !== '' && value.length !== 3) {
-    return 'CVC는 3자리여야 합니다.';
-  }
-  return '';
-};
