+++ src/constants.ts
@@ -24,3 +24,5 @@ export const CARD_BRANDS = {
 export const INITIAL_EXPIRATION = { year: { value: '', errorMessage: '' }, month: { value: '', errorMessage: '' } } as const;
 
 export const INITIAL_CVC = { value: '', errorMessage: '' };
+
+export const CARD_COMPANY = ['BC카드', '신한카드', '카카오뱅크', '현대카드', '우리카드', '롯데카드', '하나카드', '국민카드'];
